'Deleting' user in DB if he blocked the bot

+++ bot.php
@@ -147,9 +147,24 @@ $bot->onCallbackQueryData('callback_prescribe {friendId}', function (Nutgram $bo
         $puffId = (int)$trimmedPrescribe;
         $friendLang = lang($friendId);
         $inlineKeyboard = InlineKeyboardMarkup::make()->addRow(InlineKeyboardButton::make(msg('puff_decline', $friendLang), null,null, 'callback_puff_decline '.$puffId.' '.$bot->userId()),InlineKeyboardButton::make(msg('puff_approve', $friendLang), null,null, 'callback_puff_approve '.$puffId.' '.$bot->userId()))->addRow(InlineKeyboardButton::make(msg('cancel', $friendLang), null,null, 'callback_cancel'));
-        $bot->sendMessage($username.msg('prescribed_puff', $friendLang), chat_id: $friendId, reply_markup: $inlineKeyboard);
-        $bot->deleteMessage($bot->userId(),$bot->messageId());
-        $bot->sendMessage(msg('prescribe_success', lang($bot->userId())));
+        $checkStatus = checkUserStatus($friendId);
+        if ($checkStatus == 'deleted') {
+            $bot->deleteMessage($bot->userId(),$bot->messageId());
+            $bot->sendMessage(msg('prescribe_deleted_user', lang($bot->userId())));
+        } elseif ($checkStatus == 'active') {
+            try {
+                $bot->sendMessage($username.msg('prescribed_puff', $friendLang), chat_id: $friendId, reply_markup: $inlineKeyboard);
+                $bot->deleteMessage($bot->userId(),$bot->messageId());
+                $bot->sendMessage(msg('prescribe_success', lang($bot->userId())));
+            } catch (\Exception $e) {
+                if ($e->getCode() == '403') {
+                    sleep(1);
+                    UserBlockedBot($friendId);
+                    $bot->deleteMessage($bot->userId(),$bot->messageId());
+                    $bot->sendMessage(msg('prescribe_deleted_user', lang($bot->userId())));
+                }
+            }
+        }
     } elseif ($prescribe == "self") {
         $bot->deleteMessage($bot->userId(),$bot->messageId());
         $bot->sendMessage(msg('prescribe_self', lang($bot->userId())));
@@ -198,9 +213,24 @@ $bot->onCallbackQueryData('callback_puff_decline {puffId} {friendId}', function 
     createLog(TIME_NOW, $role, $bot->userId(), 'callback', 'puff_decline '.$puffId);
     $username = getUsername($bot->userId());
     updatePuff($puffId, 'decline');
-    $bot->sendMessage($username.msg('declined_puff', lang($friendId)), chat_id: $friendId);
-    $bot->deleteMessage($bot->userId(),$bot->messageId());
-    $bot->sendMessage(msg('puff_declined', lang($bot->userId())));
+    $checkStatus = checkUserStatus($friendId);
+    if ($checkStatus == 'deleted') {
+        $bot->deleteMessage($bot->userId(),$bot->messageId());
+        $bot->sendMessage(msg('puff_declined_deleted_user', lang($bot->userId())));
+    } elseif ($checkStatus == 'active') {
+        try {
+            $bot->sendMessage($username.msg('declined_puff', lang($friendId)), chat_id: $friendId);
+            $bot->deleteMessage($bot->userId(),$bot->messageId());
+            $bot->sendMessage(msg('puff_declined', lang($bot->userId())));
+        } catch (\Exception $e) {
+            if ($e->getCode() == '403') {
+                sleep(1);
+                UserBlockedBot($friendId);
+                $bot->deleteMessage($bot->userId(),$bot->messageId());
+                $bot->sendMessage(msg('puff_declined_deleted_user', lang($bot->userId())));
+            }
+        }
+    }
     $bot->answerCallbackQuery();
 });
 
@@ -212,11 +242,26 @@ $bot->onCallbackQueryData('callback_puff_approve {puffId} {friendId}', function 
     if ($status == "delay") {
         $bot->sendMessage(msg('puff_approve_delay', lang($bot->userId())));
     } else {
-        $bot->sendMessage($username.msg('approved_puff', lang($friendId)), chat_id: $friendId);
-        $bot->deleteMessage($bot->userId(),$bot->messageId());
-        $bot->sendMessage(msg('puff_approved', lang($bot->userId())));
+        $checkStatus = checkUserStatus($friendId);
+        if ($checkStatus == 'deleted') {
+            $bot->deleteMessage($bot->userId(),$bot->messageId());
+            $bot->sendMessage(msg('puff_approved_deleted_user', lang($bot->userId())));
+        } elseif ($checkStatus == 'active') {
+            try {
+                $bot->sendMessage($username.msg('approved_puff', lang($friendId)), chat_id: $friendId);
+                $bot->deleteMessage($bot->userId(),$bot->messageId());
+                $bot->sendMessage(msg('puff_approved', lang($bot->userId())));
+            } catch (\Exception $e) {
+                if ($e->getCode() == '403') {
+                    sleep(1);
+                    UserBlockedBot($friendId);
+                    $bot->deleteMessage($bot->userId(),$bot->messageId());
+                    $bot->sendMessage(msg('puff_approved_deleted_user', lang($bot->userId())));
+                }
+            }
+        }
     }
-        $bot->answerCallbackQuery();
+    $bot->answerCallbackQuery();
 });
 
 $bot->onCallbackQueryData('callback_support', function (Nutgram $bot) {
